requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-bootstrap==0.2.4
django-bootstrap4==22.1
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.8.0
django-jchart==0.4.2
django-tempus-dominus==5.1.2.17
django-tinymce==3.4.0
et-xmlfile==1.1.0
fonttools==4.33.3
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.0
pycparser==2.21
pydot==1.4.2
pydyf==0.1.2
pyparsing==3.0.8
pyphen==0.12.0
python-dateutil==1.4
pytz==2022.1
PyYAML==6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tinycss2==1.1.1
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.1
